% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-data.R
\name{build_flow_info}
\alias{build_flow_info}
\title{Build reassortment edge metadata for seqcombo}
\usage{
build_flow_info(from, to, ...)
}
\arguments{
\item{from}{source virus identifiers}

\item{to}{target virus identifiers}

\item{...}{additional columns to attach to the returned data frame}
}
\value{
a `flow_info` data frame
}
\description{
Create a `flow_info` data frame for plotting reassortment flows.
}
\author{
Guangchuang Yu
}
