% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-data.R
\name{build_segment_caption}
\alias{build_segment_caption}
\title{Build a segment caption string}
\usage{
build_segment_caption(segment_name, prefix = "Gene segments: ")
}
\arguments{
\item{segment_name}{a character vector of segment names}

\item{prefix}{prefix text used before the segment names}
}
\value{
a single caption string
}
\description{
Create a compact caption that lists segment names in order.
}
\author{
Guangchuang Yu
}
