% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-convenience.R
\name{example_seqcombo_data}
\alias{example_seqcombo_data}
\title{Example seqcombo plotting data}
\usage{
example_seqcombo_data(type = c("basic", "timeline"))
}
\arguments{
\item{type}{one of `basic` or `timeline`}
}
\value{
an object of class `seqcombo_data`
}
\description{
Return a ready-to-plot `seqcombo_data` object for demos and tests.
}
\author{
Guangchuang Yu
}
