% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{get_diff}
\alias{get_diff}
\title{access diff slot}
\usage{
get_diff(x)
}
\arguments{
\item{x}{SeqDiff object}
}
\value{
data.frame
}
\description{
access diff slot
}
\examples{
fas <- list.files(system.file("examples","GVariation", package="seqcombo"),
pattern="fas", full.names=TRUE)
x1 <- seqdiff(fas[1], reference=1)
get_diff(x1)
}

