% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/method-plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,SeqDiff,ANY-method}
\title{plot method}
\usage{
\S4method{plot}{SeqDiff,ANY}(x, width = 50, title = "auto")
}
\arguments{
\item{x}{object}

\item{width}{bin width}

\item{title}{plot title}

\item{...}{Additional argument list}
}
\value{
plot
}
\description{
plot method generics
}
\author{
guangchuang yu
}

