% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-palette.R
\name{scale_seqcombo_host}
\alias{scale_seqcombo_host}
\title{ggplot2 scales using the seqcombo host palette}
\usage{
scale_seqcombo_host(aesthetics = c("colour", "fill"), ...)
}
\arguments{
\item{aesthetics}{character vector selecting which aesthetics to supply,
any of \code{"colour"} (\code{"color"}) and \code{"fill"}}

\item{...}{additional arguments passed to
\code{\link[ggplot2]{scale_colour_manual}}}
}
\value{
a list of ggplot2 scales that can be added to a plot
}
\description{
Create paired colour/fill discrete scales wired to
[seqcombo_host_palette()], so plots colored by host metadata need no
hand-written \code{scale_color_manual()} / \code{scale_fill_manual()}
calls.
}
\examples{
data <- example_seqcombo_data()
autoplot(data, v_color = ~Host, v_fill = ~Host) +
    scale_seqcombo_host()
}
\author{
Guangchuang Yu
}
