% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-layout.R
\name{set_layout}
\alias{set_layout}
\title{set_layout}
\usage{
set_layout(
  virus_info,
  flow_info,
  layout = "layout_nicely",
  preserve_x = FALSE,
  preserve_y = FALSE
)
}
\arguments{
\item{virus_info}{virus information}

\item{flow_info}{flow information}

\item{layout}{layout method}

\item{preserve_x}{whether to preserve existing x coordinates and only update y}

\item{preserve_y}{whether to preserve existing y coordinates and only update x}
}
\value{
updated `virus_info`
}
\description{
Set layout for reassortment plot
}
\author{
Guangchuang Yu
}
