% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplot.R
\name{simplot}
\alias{simplot}
\title{simplot}
\usage{
simplot(file, query)
}
\arguments{
\item{file}{alignment fast file}

\item{query}{query sequence}
}
\value{
ggplot object
}
\description{
Sequence similarity plot
}
\examples{
fas <- system.file("examples/GVariation/sample_alignment.fa", package="seqcombo")
simplot(fas, 'CF_YL21')
}
\author{
guangchuang yu
}
