% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-convenience.R
\name{theme_seqcombo_pub}
\alias{theme_seqcombo_pub}
\title{Publication-oriented theme for seqcombo figures}
\usage{
theme_seqcombo_pub(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{base text size}

\item{base_family}{base font family}
}
\value{
a ggplot2 theme object
}
\description{
A compact ggplot2 theme tuned for reassortment diagrams.
}
\author{
Guangchuang Yu
}
