% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-convenience.R
\name{as_seqcombo_data}
\alias{as_seqcombo_data}
\title{Create a seqcombo plotting object}
\usage{
as_seqcombo_data(virus_info, flow_info = NULL)
}
\arguments{
\item{virus_info}{virus information}

\item{flow_info}{optional flow information}
}
\value{
an object of class `seqcombo_data`
}
\description{
Bundle `virus_info` and `flow_info` into a lightweight `seqcombo_data`
object for validation and autoplot methods.
}
\author{
Guangchuang Yu
}
