% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-convenience.R
\name{autoplot.seqcombo_data}
\alias{autoplot.seqcombo_data}
\title{Autoplot a seqcombo plotting object}
\usage{
\method{autoplot}{seqcombo_data}(object, ...)
}
\arguments{
\item{object}{a `seqcombo_data` object}

\item{...}{additional parameters passed to `hybrid_plot()` or `geom_genotype()`}
}
\value{
a ggplot object
}
\description{
Create a reassortment plot directly from a `seqcombo_data` object.
}
\examples{
data <- example_seqcombo_data("timeline")
ggplot2::autoplot(data, link_style = "curve")
}
\author{
Guangchuang Yu
}
