% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment-layout.R
\name{layout_timeline}
\alias{layout_timeline}
\title{layout_timeline}
\usage{
layout_timeline(
  virus_info,
  flow_info,
  time_col = "x",
  axis = c("x", "y"),
  decreasing = FALSE,
  spread = TRUE
)
}
\arguments{
\item{virus_info}{virus information}

\item{flow_info}{flow information}

\item{time_col}{column name in `virus_info` that contains temporal order}

\item{axis}{which axis should preserve time, one of `x` or `y`}

\item{decreasing}{whether to reverse the automatically computed axis}

\item{spread}{whether to evenly spread viruses sharing the same time point
along the other axis}
}
\value{
updated `virus_info`
}
\description{
Layout reassortment plots on a timeline
}
\details{
Keep temporal coordinates on one axis and automatically place viruses on the
other axis.
}
\author{
Guangchuang Yu
}
