% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot_grid}
\alias{reexports}
\alias{\%>\%}
\title{Objects exported from other packages}
\examples{
rnorm(10) \%>\% mean
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

