% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-functions.R
\docType{import}
\name{reexports}
\alias{\%>\%}
\alias{plot_grid}
\alias{reexports}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}
\keyword{internal}

