% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqcombo-package.R
\docType{package}
\name{seqcombo-package}
\alias{seqcombo}
\alias{seqcombo-package}
\title{seqcombo: Visualization Tool for Genetic Reassortment}
\description{
Provides useful functions for visualizing virus reassortment events.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/GuangchuangYu/seqcombo/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com}

}
\keyword{internal}
