% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reassortment.R
\name{set_layout}
\alias{set_layout}
\title{set_layout}
\usage{
set_layout(virus_info, flow_info, layout = "layout.auto")
}
\arguments{
\item{virus_info}{virus information}

\item{flow_info}{flow information}

\item{layout}{layout method}
}
\value{
updated virus_info
}
\description{
set layout for reassortment plot
}
\author{
guangchuang yu
}

