% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\docType{methods}
\name{show}
\alias{show}
\alias{SeqDiff-class}
\alias{show,SeqDiff-method}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{SeqDiff object}
}
\value{
message
}
\description{
show method
}
